--- tests/genai/test_labeling.py
@@ -1,6 +1,0 @@
-import pytest
-
-
-def test_labeling_raises_when_agents_not_installed():
-    with pytest.raises(ImportError, match="The `databricks-agents` package is required"):
-        import mlflow.genai.labeling  # noqa: F401
